Start session on initialization.
We're seeing the Flurry SDK not report events sometimes. This change
explicitly starts the session when we initialize the Flurry SDK which
seems to fix the issue.

This includes some build tools updates that are required to get the
tests running.

This also updates the library to consume the Flurry SDK via their jCenter() artifact.