Support v7 of Flurry SDK
Adds support for v7. The main change is how the Flurry SDK is initialized.

To test this properly, we had to introduce a factory abstraction for the
builder. This lets us test that the correct methods and arguments are
called on the builder methods.

This also changes the major version to 2.0.0-SNAPSHOT.